Move the test code into the mod.rs in order to be used by other codes.
